webapp/requirements.txt
psycopg2-binary
flask
flask-compress
requests
psycopg2
gunicorn
pyyaml
python-ldap